+++ src/main/kotlin/com/Dnight/calinify/common/colorSets/ColorSetsRepository.kt
@@ -2,4 +2,4 @@ package com.dnight.calinify.common.colorSets
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ColorSetsRepository : JpaRepository<ColorSetsEntity, Long>
+interface ColorSetsRepository : JpaRepository<ColorSetsEntity, Int>
